virtio-driver: Add Virtqueue implementation
This is the central building block for the upcoming virtio-blk driver.
It provides the functionality to enqueue requests in the virtqueue and
receive their completions.

Virtqueue is completely independent of the virtio transport being used
(such as vhost-user or vhost-vdpa), it just takes the memory for the
virtqueue from the caller, which can be anything that can be
dereferenced as a byte slice.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>