Disable feature "unsafe-op-in-unsafe-fn" by default
Do this for both the libblkio and virtio-driver crates.

This ensures that they build with the default features on any supported
rustc version. This also avoids the need to use '--no-default-features'
and 'default-features = false', which could disable other desirable
default features.

Also make src/cargo-build.sh use '--feature unsafe-op-in-unsafe-fn' so
that we continue to have the same warnings during development.

Signed-off-by: Alberto Faria <afaria@redhat.com>