Split crate 'blkio' into two
Factor out all Rust code other than src/lib.rs from the 'blkio' crate
into a separate 'blkio-rs' lib crate. This provides for a clearer
separation between the more idiomatic Rust part of the code, and the
layer translating that into a C ABI.

Further, create a workspace [1] containing all crates in the repository,
which allows us to also run rustfmt and clippy on crates 'blkio-rs' and
'virtio-driver' when building.

Finally, with IoUring now exported by a lib crate, and given it has a
new() constructor with no parameters, clippy considers that it should
implement Default, which this commit also does.

[1] https://doc.rust-lang.org/cargo/reference/workspaces.html

Signed-off-by: Alberto Faria <afaria@redhat.com>