virtio-driver: fix num_sectors when creating DiscardWriteZeroesData
`end = to_lba(offset + len)` was wrong because the last byte to be
used should be `offset + len - 1`, so `end` is pointing to the next
sector (when offset + len is aligned, which is always since we expect
this). This caused an incorrect `num_sectors`.

We can't do `to_lba(offset + len - 1)`, since we check the alignment
in `to_lba()`. So, let's set `num_sectors` directly with `to_lba(len)`.
This should be fine since we already check that `offset` is aligned.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>