LICENSE: relicense to MIT OR APACHE-2.0
The LGPL-2.1-or-later license is not widely used in Rust crates. We have
found that people often have questions about its application.

In the absence of public guidance from the Free Software Foundation,
Software Freedom Conservancy, etc it seems unlikely that there will be
clarity regarding the use of LGPL-2.1-or-later in Rust crates.

Relicense the code under the usual MIT OR Apache-2.0 dual license that
many Rust crates use. This will allow people to use and contribute to
libblkio without hesitation.

Notes:
- Also relicense the libblkio C library under MIT OR Apache-2.0 even
  though LGPL-2.1-or-later would be more readily accepted in non-Rust
  applications. This keeps licensing more consistent and simple.
- Use parentheses "(MIT OR Apache-2.0)" in license headers because this
  is widely done although the SPDX grammar does not require it.
- Do not use parentheses in Config.toml because crates.io did not
  support them until 2021 and they are rare.

Signed-off-by: Alberto Faria <afaria@redhat.com>
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>
Signed-off-by: Kevin Wolf <kwolf@redhat.com>
Signed-off-by: German Maglione <gmaglione@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>