blkio: Add a MemoryRegionRegistry helper struct
Move generic memory region state into it, and make it responsible for
freeing allocated memory regions when dropped. Document the guarantee
that mapped memory regions are implicitly dropped when drivers are
dropped.

Append a `MemoryRegionRegistry` field to `Blkio`. Since this field is
guaranteed to be dropped after the `Box<dyn Driver>` field, explicitly
unmapping regions in `MemoryRegionRegistry::drop` is not necessary.

Signed-off-by: Alberto Faria <afaria@redhat.com>