virtio-driver/vhost_vdpa: fix virtqueue allocation to work with VDUSE
VDUSE requires that memory should be allocated with an associated fd,
to be shared with the device emulated in the VDUSE application, as
Yongji Xie confirmed here:
  https://lore.kernel.org/qemu-devel/CACycT3vMVbm94OYtUgB9bS6_pU6FKh1Y6kWPyAxi4rvjUfQ3qg@mail.gmail.com/

Let's use the same approach as vhost-user, allocating the virtqueues
memory through a memory mapped anonymous file.
We also set `needs_mem_region_fd` to true; this is only required for
VDUSE devices, but so far we don't know how to distinguish it, so
better to be conservative.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>