virtio-driver: avoid making virtio_blk and device modules public
Commit c04cf99 ("virtio-driver/virtio_blk: read num_queues from
config space") added `virtio_driver::devices::virtio_blk::max_queues`,
making devices and virtio_blk modules public.

This mixed up the current approach we have for handling namespaces in
virtio_driver, so let's revert the change, rename the function to
`virtio_blk_max_queues()`, and expose it as we do for other functions
(e.g. `validate_lba()`).

We'll try a more idiomatic approach in the future, but it will require
breaking the API.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>