virtio-driver/transports/vhost_user: Avoid constructing unaligned reference
The newly-released Rust 1.69 disallows creating unaligned references,
which we do in `VhostUserMsg::recv` to construct a pointer to the
unaligned `VhostUserMsg::payload` field. We would ideally use
`std::ptr::addr_of` instead, but it isn't available in Rust 1.48, which
is the version packaged by stable Debian.

Work around this by instead obtaining a pointer to the `VhostUserMsg`
and advancing it by the size of `VhostUserHeader`, which is also the
offset at which `VhostUserMsg::payload` starts.

Reported-by: Nir Soffer <nsoffer@redhat.com>
Signed-off-by: Alberto Faria <afaria@redhat.com>