Raise minimum Rust version to 1.56
This is in preparation to update our dependencies to their latest
versions, which will facilitate libblkio packaging for Debian bookworm.

This makes libblkio incompatible with Debian bullseye's packaged Rust
version (1.48), but it is more important to package libblkio for the
next version of Debian (bookworm), which will be released in a few
months.

Also update our crates to the Rust 2021 edition, which was introduced
precisely with Rust 1.56.

Signed-off-by: Alberto Faria <afaria@redhat.com>