virtio-driver: Fix used_notif_enabled initialization
Current code initialize used_notif_enabled to true, so when EVENT_IDX is
enabled and we use polling, we store the next_idx in used_event instead
of next_idx - 1, using used buffers notifications instead of suppressing
them.

Testing show no difference in performance, since we do the same work in
both cases, but it prevents unneeded notifications in the device.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>