virtio-driver: fix used_event memory ordering
The EVENT_IDX mechanism relies on the following memory ordering in the
driver for Used Buffer Notification Suppression:

1. Store used_event to re-enable notifications.
2. Load used.idx to check if a used buffer snuck in.

The device follows the opposite ordering:

1. Store used.idx so the driver can see the used buffer.
2. Load used_event to determine whether to send a notification.

This ensures that used buffers are never missed by the driver.

The driver currently does not implement this ordering correctly:

1. fn update_used_event(&self) {
       ...(*self.avail.event).store(..., Ordering::Release)...
   }
2. fn load_next_idx(&self) -> u16 {
       ...(*self.ptr).idx.load(Ordering::Acquire)...
   }

According to the nomicon, Ordering::Release "ensures that every access
before it stays before it" and Ordering::Acquire "ensures that every
access after it stays after it".

Nothing ensures that the store happens before the load. This can lead to
hangs where the driver does not detect used buffers because the device
does not send a notification.

The hangs occur when building with --buildtype=debugoptimized and
running:

  build/examples/blkio-bench --iodepth 2 --count 1000000 virtio-blk-vhost-user path=/tmp/vhost-user-blk.sock

Thanks to Nir Soffer for debugging this with me.

Co-authored-by: Nir Soffer <nsoffer@redhat.com>
Reported-by: Nir Soffer <nsoffer@redhat.com>
Cc: Hanwen Gu <allezvicki@gmail.com>
Cc: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>