virtio-driver: Never update used_event if not needed
When used buffers notification are suppressed (used_notif_enabled=false),
never update used_event instead of updating it for every completion.

With this not all notifications are suppressed and the device will
trigger one notification when the used.idx wraps, but we save atomic
store and fence for every completion.

Testing shows only tiny improvement in kiops, but it is clear that the
cost of calling blkioq_do_io() is much smaller now, and we do 1.32 times
more calls in the same time. This may help if we disable used buffer
notifications without polling.

Before:

    $ build/examples/blkio-bench --iodepth 64 --runtime 60 virtio-blk-vhost-user \
        --poll path=/tmp/vhost-user-blk.sock | jq .aggregate
    {
      "kiops": 1785.246,
      "min_lat_us": 4.573,
      "mean_lat_us": 35.849,
      "max_lat_us": 156.085,
      "calls": 705888046,
      "min_completions": 1.00,
      "mean_completions": 0.15,
      "max_completions": 64.00
    }

After:

    $ sleep 60; build/examples/blkio-bench --iodepth 64 --runtime 60 --poll \
        virtio-blk-vhost-user path=/tmp/vhost-user-blk.sock | jq .aggregate
    {
      "kiops": 1812.050,
      "min_lat_us": 6.879,
      "mean_lat_us": 35.319,
      "max_lat_us": 157.409,
      "calls": 935185666,
      "min_completions": 1.00,
      "mean_completions": 0.12,
      "max_completions": 52.00
  }

Signed-off-by: Nir Soffer <nsoffer@redhat.com>