virtio-driver: fix avail_event memory ordering
The EVENT_IDX mechanism relies on the following memory ordering in the
driver for Available Buffer Notification Suppression:

1. Store avail.idx.
2. Load avail_event to check if an Available Buffer Notification is
   required.

The device follows the opposite ordering:

1. Store avail_event to re-enable Available Buffer Notifications.
2. Load avail.idx to check if an available buffer snuck in.

This ensures that available buffers are never missed by the device. The
VIRTIO spec mentions memory barriers in Step 6 here:
https://docs.oasis-open.org/virtio/virtio/v1.2/cs01/virtio-v1.2-cs01.html#x1-6400013

The driver currently does not implement this ordering correctly:

1. fn store_next_idx(&self) {
       ...(*self.ptr).idx.store(self.next_idx.0.to_le(),
                                Ordering::Release)...
   }
2. fn load_event(&self) -> u16 {
       ...(*self.event).load(Ordering::Acquire)...
   }

According to the nomicon, Ordering::Release "ensures that every access
before it stays before it" and Ordering::Acquire "ensures that every
access after it stays after it".

Nothing ensures that the store happens before the load. This can lead to
hangs where the device does not detect available buffers because the
driver does not send a notification.

Thanks to Nir Soffer for debugging this with me.

Co-authored-by: Nir Soffer <nsoffer@redhat.com>
Reported-by: Nir Soffer <nsoffer@redhat.com>
Cc: Hanwen Gu <allezvicki@gmail.com>
Cc: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>