virtio-driver/virtqueue: support multiple virtqueue format
Add `VirtqueueFormat` trait to support both split and packed virtqueues.
Some of the shared code is left in Virtqueue and the specific code moved
to VirtqueueSplit.

Change the nomenclature a bit to be compatible with split and packed
virtqueue.

The major difference is how to identify a chain of descriptors: for split
is used the index of the first descriptor, while for packed is a generic
identifier.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>